Corfudb version: 0.3.0.20200204063830.3213

List of fixes:
CorfuStore: Add epoch, CLEAR to streaming API
CorfuStore: fix StreamListner's hashCode() for recognition in Java collections
Log stream Id when trimmed address is encountered by clients.
Remove AbstractCorfuWrapper (#2316)
Explicitly Close Map Streams
VersionLockedObject and CorfuTable Deadlock (#2325)
Do not log empty secondary indexes when creating CorfuTable.
Make benchmark module to work with latest version of universe fr (#2334)
PersistedStreamingMap Iterator Bug (#2317)
Make VmManager more reliable in terms of catching runtime exceptions from ViJava lib (#2332)
Ignore flattened pom files (#2333)
Change checkpoint initialization logic (#2307)
Add DiskBackedCorfuClientTest (#2238)
Change logSizeQuotaPercentage from int to double (#2324)
Update dropwizard-metrics to 4.1.2 (#2319)
Remove Heart Beat Executor (#2267)
Mask TrimmedException in CorfuCompileProxy (#2323)
End-to-end compression: remove compressed flag and related checks.
Update assertj to 3.14.0 (#2318)
Improved Thread Management (#2272)
Some code cleanup. Remove the not need args for checkpoint funciton. (#2303)
Update slf4j logger to 1.7.25 (#2314)
CorfuBrowser: Use Nullable, refactor code, tests

Issue: #2497135
Issue: #2497127

+++ it/src/main/java/org/corfudb/universe/node/server/vm/VmCorfuServerParams.java
@@ -30,7 +30,7 @@ public class VmCorfuServerParams extends CorfuServerParams {
     public VmCorfuServerParams(
             VmName vmName, int port, Mode mode, Persistence persistence,
             Level logLevel, String clusterName, Duration stopTimeout, String serverVersion,
-            Path serverJarDirectory, String dockerImage, int logSizeQuotaPercentage) {
+            Path serverJarDirectory, String dockerImage, double logSizeQuotaPercentage) {
 
         super(
                 port, mode, persistence, logLevel, clusterName, stopTimeout,
